test(payment): test에 필요한 mock server 생성

+++ payment/src/main/java/com/devticket/payment/payment/infrastructure/external/dto/TossPaymentConfirmResponse.java
@@ -5,7 +5,7 @@ public record TossPaymentConfirmResponse(
     String orderId,
     String method,
     String status,
-    Long totalAmount,
+    Integer totalAmount,
     String approvedAt
 ) {
 }
